Improved batch extractor
Added "Extract all..." context menu when right clicking empty space in file tree view
Removed FBX SDK

+++ src/MGF/AssetManager.h
@@ -12,7 +12,7 @@ namespace MGF {
 		MGF::Asset::AssetPtr Get(const MGF::File& file);
 
 	private:
-		std::unordered_map<std::uint64_t, MGF::Asset::AssetPtr> Assets;
+		std::unordered_map<std::uint32_t, MGF::Asset::AssetPtr> Assets;
 	};
 
 }
